// types/tables.ts
export type WebWidgetData = {
    BranchID: number;
    ReportID: number;
    reportValue1: string;    // SubeAdi
    reportValue2: number;    // TC (Cari dönem ciro)
    reportValue3: number;    // GHTC (Geçen hafta aynı saat ciro)
    reportValue4: number;    // GHTCTUM (Geçen hafta tüm gün ciro)
    reportValue5: number;    // KisiSayisi
    reportValue6: number;    // GHKisiSayisi
    reportValue7: number;    // GHKisiSayisiTUM
    reportValue8: number;    // Oran
    reportValue9: number;    // GecenHaftaOran
}